Add subscription checking to Post Proof button on Events page
