+++ imports/ui/components/SkillTrees/Events/JoinEventButton.jsx
@@ -2,26 +2,9 @@ import { Meteor } from 'meteor/meteor';
 import { useFind, useSubscribe } from 'meteor/react-meteor-data/suspense';
 import React from 'react';
 import { EventCollection } from '/imports/api/collections/Events';
-import { SubscriptionsCollection } from '/imports/api/collections/Subscriptions';
 
-export const JoinEventButton = ({ eventId, skillTreeId }) => {
+export const JoinEventButton = ({ eventId, skillTreeId, disabled }) => {
   const userId = Meteor.userId();
-
-  useSubscribe('subscriptions');
-  const subscription = useFind(
-    SubscriptionsCollection,
-    [
-      {
-        userId: { $eq: userId },
-        skillTreeId: { $eq: skillTreeId },
-        active: { $eq: true }
-      },
-      { fields: { _id: 1 } }
-    ],
-    [userId, skillTreeId]
-  )[0];
-
-  const isSubscribed = !!subscription;
 
   useSubscribe('events');
   const checkJoined =
@@ -33,7 +16,9 @@ export const JoinEventButton = ({ eventId, skillTreeId }) => {
   const isJoined = !!checkJoined;
 
   const denyJoin = () => {
-    alert('You must be subscribed to the skilltree to join the event.');
+    alert(
+      'You must be subscribed to the skilltree to join or leave the event.'
+    );
   };
 
   /**
@@ -63,11 +48,11 @@ export const JoinEventButton = ({ eventId, skillTreeId }) => {
   return (
     <button
       className="block py-2 pl-3 pr-4 md:p-0 cursor-pointer"
-      onClick={!isSubscribed ? denyJoin : isJoined ? leaveEvent : joinEvent}
+      onClick={disabled ? denyJoin : isJoined ? leaveEvent : joinEvent}
     >
       <div
         className={`px-3 py-2 rounded ${
-          !isSubscribed
+          disabled
             ? 'w-full sm:w-auto bg-gray-400 text-gray-200 font-semibold py-2 px-4 rounded-lg shadow hover:bg-[#2a7d60] transition-colors'
             : isJoined
               ? 'w-full sm:w-auto bg-white text-black font-semibold py-2 px-4 rounded-lg hover:bg-gray-200 transition-colors'
