+++ imports/ui/pages/RankedEvent.jsx
@@ -1,19 +1,19 @@
-import React, { Suspense, useEffect, useState } from 'react';
-import { Helmet } from 'react-helmet';
 import { Meteor } from 'meteor/meteor';
+import React, { Suspense, useState } from 'react';
+import { Helmet } from 'react-helmet';
 
 // JSX UI
-import { useSubscribe, useFind } from 'meteor/react-meteor-data/suspense';
+import { useFind, useSubscribe } from 'meteor/react-meteor-data/suspense';
 import { useParams } from 'react-router-dom';
-import { SkillTreeCollection } from '/imports/api/collections/SkillTree';
 import { EventCard } from '../components/RankedEvents/EventCard';
-import { NavigationMenu } from '../components/SkillTrees/NavigationMenu';
 import { EventInfoModal } from '../components/RankedEvents/EventInfoModal';
+import { NavigationMenu } from '../components/SkillTrees/NavigationMenu';
+import { SkillTreeCollection } from '/imports/api/collections/SkillTree';
 
-import { SubscriptionsCollection } from '/imports/api/collections/Subscriptions';
 import { JoinEventButton } from '../components/SkillTrees/Events/JoinEventButton';
-import { EventCollection } from '/imports/api/collections/Events';
 import { ProofUploadButton } from '../components/SkillTrees/Skill/ProofUploadButton';
+import { EventCollection } from '/imports/api/collections/Events';
+import { SubscriptionsCollection } from '/imports/api/collections/Subscriptions';
 
 export const RankedEvent = () => {
   const { skilltreeId } = useParams();
@@ -69,6 +69,22 @@ export const RankedEvent = () => {
 
   // Filter state: 'default' or 'upvotes'
   const [filter, setFilter] = useState('default');
+
+  useSubscribe('subscriptions');
+  const subscription = useFind(
+    SubscriptionsCollection,
+    [
+      {
+        userId: { $eq: userId },
+        skillTreeId: { $eq: skilltreeId },
+        active: { $eq: true }
+      },
+      { fields: { _id: 1 } }
+    ],
+    [userId, skilltreeId]
+  )[0];
+
+  const isUserSubscribed = !!subscription;
 
   if (!skilltree) return <div>Skill Tree not found</div>;
 
@@ -132,9 +148,14 @@ export const RankedEvent = () => {
               onUploadProof={() => {
                 console.log('onUploadProof');
               }}
+              disabled={!isUserSubscribed}
             />
 
-            <JoinEventButton eventId={eventId} skillTreeId={skilltreeId} />
+            <JoinEventButton
+              eventId={eventId}
+              skillTreeId={skilltreeId}
+              disabled={!isUserSubscribed}
+            />
             <button
               onClick={() => setIsModalOpen(true)}
               className="w-full sm:w-auto bg-[#328E6E] text-white font-semibold py-2 px-4 rounded-lg shadow hover:bg-[#2a7d60] transition-colors"
